Changing behavior if date of remote sha1 is not present
There are two cases in which the date of the remote sha1 file cannot be
determined.

 1. init action: there is no remote sha1 file yet. We don't want to
overwrite remote files if they are already there. The detection of
remote changes aborts here as well now.

 2. server doesn't provide dates: this makes the detection of remote
changes via date checks impossible. Nevertheless, the program should
abort before overwriting files which's state is unclear. If it's a
temporary problem, the user can try it again. If it is a permanent
problem, the user has to deactivate the detection of remote changes and
work without checks. A full file diff is very expensive, but can be
performed manually with 'git-ftp pull'.

These changes were made:
 * Previously, if the server didn't provide a date,
get_file_last_modiefied() returned the ouput of `date` with an empty
string which is 12am at the current day. Now it returns nothing.
 * remote_file_modified() tests if the date of the remote sha1 file is
present. If it is not present, all found timestamps are seen as newer.
 * Remote file modifications are also checked if no date of the remote
sha1 file is present (e.g. during init).
 * The abortion of init is covered by a unit test.