Testing an empty URL instead of an empty password
This has two main advantages:

1.
If you use an anonymous FTP server for testing, the login with an empty
password succeeds and this test fails.

2.
An empty URL results in an earlier error message that doesn't involve a
network connection. This saves 20 seconds (invalid password timeout) in
my testing setup.