In test suite: always appending a slash to root path
Invoking the test suite like this caused errors:
  GIT_FTP_USER=ftp GIT_FTP_PASSWD=ftp GIT_FTP_ROOT=localhost/test ./tests/git-ftp-test.sh
The GIT_FTP_ROOT required a trailing slash. That is changed by this
commit.

That slash isn't required any more. If it's added, the path will contain
two slashes which isn't a problem. I don't see a point in adding code
for canonicalising the path in the test suite.