Checking remote access before initial upload
The init action tries to upload a heap of files before it can check
curl's exit status and realise that the upload doesn't work. That can
take a while.

An additional check was added. It just tries to list the target
directory with the --ftp-create-dirs switch. If that fails, something
is wrong.