Supporting regex for ignore file
The patterns in .git-ftp-ignore where applied to the output of `git
diff` or `git ls-files`:
    M modifield_file.txt
    D deleted_file.txt

You had to consider that to ignore a single file (issue #80).

Now that output is split into a list of files to upload and another list
of files to delete. These lists can be filtered properly.

Conflicts:
	git-ftp