Separate variable CURL_PROTOCOL for ftpes
If REMOTE_PROTOCOL was set to ftpes, the curl option --ssl was set and
the protocol was reset to ftp. That changed variable was passed onto
submodules which used ftp instead of ftpes.

This commit add a new variable CURL_URL that is said to ftp while
REMOTE_PROTOCOL remains ftpes.

Another test case was added to test submodules in conjunction with
--syncroot.

Other protocols than ftp remain untested.