insert space to not mess up lftp exclude options
the last line of .git-ftp-ignore as ecclude and var ignore in function download_remote_updates get attached to closely, so that something like --exclude XYZ--exclude ^.git/ happens. The last line and the first ignore statement thereby don't work. Inserting a space at the beginning of the ignore variable solves this problem.