fix: error handling with new sftp directory creation
Pull request #504 introduced new logic to create directories with sftp.
As mentioned in the discussion to this (https://github.com/git-ftp/git-ftp/pull/504#discussion_r281009161) the error handling is not perfect. This commit fixes this issue.

At first do the curl request as before bot cache the exit code in a local variable.
The we can check if sftp was used and `$EXIT_CODE` was `78`. If so do the creation stuff. If not check the $EXIT_CODE for other errors.