fix: submodule handling
The check if a file/directory is a submodule was broken. A detailed analysis to this issue can be found [here](https://github.com/git-ftp/git-ftp/issues/295#issuecomment-488627926).

This fixes this issue by setting `SYNCROOT` earlier in the call stack, so `cache_git_submodules` will work as expected.

fixes #295

**Limitations**
To the current state, handling of submodules only works with username and password. Usage of ssh-keys with submodules does not work. (see #507)