enhancement: Add support for key auth with submodules
Adds option to use ssh-keys in combination with submodules.

So far user and passwd were set at the beginning of `set_submodule_args()`, even if passwd was not set.
This commit adds a check if sftp is used with ssh-keys and - if so - it adds them instead of passwd. If sftp is used with passwd or any other protocol, passwd is set as before.

closes #507