Add support for --insecure for download and pull
This adds a new handling of the lftp command used during download, pull and snapshot action.
The --insecure flag was ignored on this commands. This adds the ssl:verify-certificate to the lftp command to fix this.
A detailed description to the problem can be found here: https://github.com/git-ftp/git-ftp/issues/366#issuecomment-490266930

In addition to this I tried to make this part of the code mor readable.