Documentation: Add description for macOS keychain
I had some truble to getting the keychain to work with the old description so I decided to add a more detailed one.
I renamed user to account and account to host so it fits the naming Apple is using in their tools.

Is it to extensive?

In addition made some small changes to the code to make it a litte more readable.

additin to #389