Check curl exit code during action catchup
Add the test for curl exit code to action catchup the same way.
The check is done in `upload_local_sha1()`.

Since `upload_file()` is only used twice (in `remote_lock()` and in `upload_local_sha1()`) and the result is already checked after the call, I decided to remove the call to `check_exit_status` inside of `upload_file`. So the calling function can decide how to handle the different errors and exit codes. With the call inside it's not possible to differ between different exit codes of the curl command from outside the function.

Also add a test case for this.