Enhancement: Process boolean configs as integers instead of strings
Convert the type of `CURL_INSECURE` and `CURL_DISABLE_EPSV` to integers.
This makes it more intuitive/readable and easier to add other defaults for booleans later.

Also rename `CURL_INSECURE` to `INSECURE` since it is used for LFTP as well.

And add tests for false values for boolean configs.