Feature: Add new option --auto-stash and associated config
The auto-stash feature calls `git stash` before checking for a dirty repo.
It also stashes untracked/new files so the repo is really clean.
After doing the other stuff the stash is applied (popped) again,
so the state of the working tree is restored.

If there are no files to stash, git stash succeds without adding a stash entry.
Therefore we have to check if the last stash is really made from Git-ftp before poping.

In addition to the option the config is suported, too.