Fix: Give correct error message for unrecognised options
The first unknown thing was always treated as the URL. So in cases where the unrecognised option was before the URL the URL caused the error.
This fixes this by checking the unknown thing if it could be meant as a option. If it is a option it is not set as the URL but produces the error.

In addition two tests are added to check the produced error messages in both cases.

fixes #522