Clarifying setup of custom curl for Apple Silicon
Hi! Newer versions of macs with Apple Silicon don't seem to come built in with libpsl, so I've added this line to make sure people know to install it up-front when going through these instructions. 

Also, Apple Silicon has a different folder prefix for HomeBrew, so using the brew --prefix command to build the command fixes the issue for those that are migrating from an Intel mac to an Apple Silicon one.